refactor(util): tsc-util を他の util と統合
