+++ src/usecase/watchMetrics.ts
@@ -1,7 +1,7 @@
 import { OptionValues } from '../setting/model';
 import chokidar from 'chokidar';
 import { pipe, piped, tap } from 'remeda';
-import { isTsFile } from '../tsc-utils';
+import { isTsFile } from '../utils/tsc-util';
 import { Table } from 'console-table-printer';
 import chalk from 'chalk';
 import { CodeMetrics, Score } from '../feature/metric/metricsModels';
